Scripts to start and stop the server
start_server.sh will start a new droplet and terraria server on
DigitalOcean; (it does this by starting them all in a docker
container)

stop_server.sh will stop the server, make a backup of the gameworld
file and destroy the droplet (this will help prevent billing).

the stuff in /config .. I'm not sure about. Am keeping them for now.